ecommerce-backend/gateway/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
requests==2.32.3
PyJWT==2.8.0
